Decompose GitHub workflow modules
