+++ senpai_agent/github_tools/__init__.py
@@ -1,0 +1,75 @@
+"""Role-scoped GitHub tools with one schema per workflow action."""
+
+from .contracts import (
+    AcceptResultOnCurrentBaseAction,
+    AssignmentVersion,
+    CloseExperimentAction,
+    CreateAssignmentAction,
+    GitHubMutationObservation,
+    MergeExperimentAction,
+    PublishAdvisorBranchAction,
+    RepairAssignmentRoutingAction,
+    RequestAssignmentRevisionAction,
+    RespondToHumanIssueAction,
+    SendAssignmentFeedbackAction,
+    SubmitExperimentResultAction,
+)
+from .definitions import (
+    AcceptResultOnCurrentBaseTool,
+    CloseExperimentTool,
+    CreateAssignmentTool,
+    MergeExperimentTool,
+    PublishAdvisorBranchTool,
+    RepairAssignmentRoutingTool,
+    RequestAssignmentRevisionTool,
+    RespondToHumanIssueTool,
+    SendAssignmentFeedbackTool,
+    SubmitExperimentResultTool,
+)
+from .pull_requests import (
+    GetPRsAction,
+    GetPRsObservation,
+    GetPRsTool,
+    PRManifestObservation,
+)
+from .runtime import (
+    GitHubCredentials,
+    GitHubToolRuntime,
+    clear_github_credentials,
+    configure_github_credentials,
+)
+from .toolset import GitHubWorkflowToolSet
+
+__all__ = (
+    "AcceptResultOnCurrentBaseAction",
+    "AcceptResultOnCurrentBaseTool",
+    "AssignmentVersion",
+    "CloseExperimentAction",
+    "CloseExperimentTool",
+    "CreateAssignmentAction",
+    "CreateAssignmentTool",
+    "GetPRsAction",
+    "GetPRsObservation",
+    "GetPRsTool",
+    "GitHubCredentials",
+    "GitHubMutationObservation",
+    "GitHubToolRuntime",
+    "GitHubWorkflowToolSet",
+    "MergeExperimentAction",
+    "MergeExperimentTool",
+    "PRManifestObservation",
+    "PublishAdvisorBranchAction",
+    "PublishAdvisorBranchTool",
+    "RepairAssignmentRoutingAction",
+    "RepairAssignmentRoutingTool",
+    "RequestAssignmentRevisionAction",
+    "RequestAssignmentRevisionTool",
+    "RespondToHumanIssueAction",
+    "RespondToHumanIssueTool",
+    "SendAssignmentFeedbackAction",
+    "SendAssignmentFeedbackTool",
+    "SubmitExperimentResultAction",
+    "SubmitExperimentResultTool",
+    "clear_github_credentials",
+    "configure_github_credentials",
+)
