+++ senpai_agent/github_workflow/text.py
@@ -1,0 +1,61 @@
+"""Render and update human-readable workflow comments."""
+
+import re
+from typing import Literal
+
+from senpai_agent.github_workflow.errors import WorkflowPreconditionError
+from senpai_agent.models import AssignmentRecord, render_assignment_marker
+
+
+_ROLE_COMMENT_PREFIX = re.compile(r"^(?:ADVISOR|STUDENT(?: [^:\s]+)?):[ \t]*")
+
+
+def marker_body(marker: str, content: str) -> str:
+    content = content.strip()
+    if not content:
+        raise ValueError("marker comment content must not be empty")
+    body = f"{marker}\n\n{content}"
+    _validate_marker(marker, body)
+    return body
+
+
+def role_prefixed_comment(
+    body: str,
+    role: Literal["advisor", "student"],
+) -> str:
+    marker, separator, content = body.partition("\n\n")
+    if not (separator and marker.startswith("<!-- senpai-")):
+        marker, separator, content = "", "", body
+    content = _ROLE_COMMENT_PREFIX.sub("", content)
+    return f"{marker}{separator}{role.upper()}: {content}"
+
+
+def replace_assignment_marker(
+    body: str,
+    assignment: AssignmentRecord,
+) -> str:
+    replacement = render_assignment_marker(assignment)
+    lines = body.splitlines()
+    indexes = [
+        index
+        for index, line in enumerate(lines)
+        if line.startswith("<!-- senpai-assignment:")
+    ]
+    if len(indexes) != 1:
+        raise WorkflowPreconditionError(
+            "pull request must contain exactly one assignment marker"
+        )
+    lines[indexes[0]] = replacement
+    return "\n".join(lines)
+
+
+def _validate_marker(marker: str, body: str) -> None:
+    if (
+        "\n" in marker
+        or "\r" in marker
+        or not marker.startswith("<!-- senpai-")
+        or not marker.endswith("-->")
+    ):
+        raise ValueError("marker must be one hidden Senpai marker")
+    if body.splitlines().count(marker) != 1:
+        raise ValueError("comment body must contain the marker exactly once")
